Avoid stutter when removing name search
This avoids stopping and zooming out when a matched name is removed from
the input and garbage text is quickly inserted again. Instead, if one
quickly inserts a correct text again after emptying, the transition ends
at the current zoom level and starts zooming again, as well as vice
versa (removing text quickly after starting a zoom to the match).