Timelapse: Fix 'missing' node/link errors
If a link is recently added in the timelapse, then its properties are
not yet updated according to the force simulation. Most importantly, its
source and target properties are strings rather than references to the
nodes. This means that checks such as whether a link refers to a new
project, or whether a node has no more other links may fail. The force
simulation then detects that there are missing nodes and errors out.

Also clean up the current projects check to add all new projects even if
there are some existing ones.